chore: Stop spamming logs with JSON blobs
Collection rules and collection configuration are being logged to the
file, and it makes it hard to investigate issues with them taking over
most of the content of the log.

The code being run should contain enough information to figure out what
what the important parts of the configuration were.